# docs/architecture/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecordingCompleteRequest": {
        "properties": {
          "asset_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Url"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          }
        },
        "title": "RecordingCompleteRequest",
        "type": "object"
      },
      "RecordingFailRequest": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "RecordingFailRequest",
        "type": "object"
      },
      "RecordingSession": {
        "properties": {
          "asset_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "recording_id": {
            "title": "Recording Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "source": {
            "default": "primary",
            "title": "Source",
            "type": "string"
          },
          "state": {
            "default": "recording",
            "enum": [
              "recording",
              "stopping",
              "uploading",
              "completed",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "target_dir": {
            "default": "ingest/live",
            "title": "Target Dir",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "recording_id",
          "session_id",
          "node_id",
          "project"
        ],
        "title": "RecordingSession",
        "type": "object"
      },
...
      "RecordingSessionResponse": {
        "properties": {
          "recording": {
            "$ref": "#/components/schemas/RecordingSession"
          }
        },
        "required": [
          "recording"
        ],
        "title": "RecordingSessionResponse",
        "type": "object"
      },
      "RecordingStartRequest": {
        "properties": {
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "recording_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recording Id"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "source": {
            "default": "primary",
            "title": "Source",
            "type": "string"
          },
          "target_dir": {
            "default": "ingest/live",
            "title": "Target Dir",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "node_id",
          "project"
        ],
        "title": "RecordingStartRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "media-sync-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/recordings/start": {
      "post": {
        "operationId": "start_recording_session_api_recordings_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Recording Session",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/recordings/{recording_id}": {
      "delete": {
        "operationId": "delete_recording_session_api_recordings__recording_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Recording Session",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/recordings/{recording_id}/complete": {
      "post": {
        "operationId": "complete_recording_session_api_recordings__recording_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Recording Session",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/recordings/{recording_id}/fail": {
      "post": {
        "operationId": "fail_recording_session_api_recordings__recording_id__fail_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingFailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fail Recording Session",
        "tags": [
          "recordings"
        ]
      }
    },
...
  }
}